Handle Firefox 131 using the full profile path in X properties
As of Firefox 131 or so, Firefox has switched to putting the full path
to the profile in its X properties, instead of the profile name. We
recognize this and handle the new format (as well as the old one),
allowing users to provide either a full path or the old profile name.